requirements.txt
flask==3.0.0
pymongo==4.5.0
python-dotenv==1.0.0
boto3==1.28.44
openai==1.3.0
langchain-openai==0.0.2
langchain-core==0.1.12
llama-index==0.9.27
gunicorn==21.2.0